Extract saving operations to separate module
