+++ src/lsdb/catalog/catalog.py
@@ -1,30 +1,19 @@
 from __future__ import annotations
 
 import dataclasses
-import os
-from copy import copy
-from importlib.metadata import version
 from typing import Any, Dict, List, Tuple, Type, Union, cast
 
 import dask.dataframe as dd
 import hipscat as hc
-from hipscat.io import FilePointer, file_io
 from hipscat.pixel_math import HealpixPixel
-from typing_extensions import TypeAlias
 
+from lsdb import io
 from lsdb.catalog.dataset.dataset import Dataset
 from lsdb.core.cone_search import cone_filter
 from lsdb.core.crossmatch.abstract_crossmatch_algorithm import AbstractCrossmatchAlgorithm
 from lsdb.core.crossmatch.crossmatch_algorithms import BuiltInCrossmatchAlgorithm
 from lsdb.dask.crossmatch_catalog_data import crossmatch_catalog_data
-from lsdb.io.file_io import write_dataframe_to_parquet
-
-DaskDFPixelMap = Dict[HealpixPixel, int]
-
-# Compute pixel map returns a tuple. The first element is
-# the number of data points within the HEALPix pixel, the
-# second element is the list of pixels it contains.
-HealpixInfo: TypeAlias = Tuple[int, List[int]]
+from lsdb.types import DaskDFPixelMap
 
 
 # pylint: disable=R0903, W0212
@@ -36,7 +25,6 @@ class Catalog(Dataset):
         hc_structure: `hipscat.Catalog` object representing the structure
                       and metadata of the HiPSCat catalog
     """
-
     hc_structure: hc.catalog.Catalog
 
     def __init__(
@@ -96,22 +84,6 @@ class Catalog(Dataset):
             raise ValueError(f"Pixel at order {order} pixel {pixel} not in Catalog")
         partition_index = self._ddf_pixel_map[hp_pixel]
         return partition_index
-
-    def _get_partition_info_dict(
-        self, ddf_points_map: Dict[HealpixPixel, int]
-    ) -> Dict[HealpixPixel, HealpixInfo]:
-        """Creates the partition info dictionary
-
-        Args:
-            ddf_points_map (Dict[HealpixPix,int]): Dictionary mapping each HealpixPixel
-                to the respective number of points inside its partition
-
-        Returns:
-            A partition info dictionary, where the keys are the HEALPix pixels and
-            the values are pairs where the first element is the number of points
-            inside the pixel, and the second is the list of destination pixel numbers.
-        """
-        return {pixel: (length, [pixel.pixel]) for pixel, length in ddf_points_map.items()}
 
     @property
     def name(self):
@@ -275,49 +247,6 @@ class Catalog(Dataset):
         ddf_partition_map = {pixel: i for i, pixel in enumerate(pixels_in_cone)}
         return Catalog(cone_search_ddf, ddf_partition_map, filtered_hc_structure)
 
-    def write_partitions(
-        self, base_catalog_dir_fp: FilePointer, storage_options: Union[Dict[Any, Any], None] = None
-    ) -> Dict[HealpixPixel, int]:
-        """Saves catalog partitions as parquet to disk
-
-        Args:
-            base_catalog_dir_fp (FilePointer): Path to the base directory of the catalog
-            storage_options (dict): dictionary that contains abstract filesystem credentials
-
-        Returns:
-            A dictionary mapping each HEALPix pixel to the number of data points in it.
-        """
-        pixel_to_partition_size_map = {}
-        for pixel, partition_index in self._ddf_pixel_map.items():
-            partition = self._ddf.partitions[partition_index].compute()
-            pixel_path = hc.io.paths.pixel_catalog_file(base_catalog_dir_fp, pixel.order, pixel.pixel)
-            write_dataframe_to_parquet(partition, pixel_path, storage_options)
-            pixel_to_partition_size_map[pixel] = len(partition)
-        return pixel_to_partition_size_map
-
-    def create_modified_hc_structure(self, catalog_base_dir: str, catalog_name: str, **kwargs):
-        """Creates a modified version of the HiPSCat catalog structure
-
-        Args:
-            catalog_base_dir (str): Base location for the catalog
-            catalog_name (str): The name of the catalog to be saved
-            **kwargs: The remaining parameters to be updated in the catalog info object
-
-        Returns:
-            A HiPSCat structure, modified with the parameters provided.
-        """
-        new_hc_structure = copy(self.hc_structure)
-        # Update catalog path and name in the structure
-        new_hc_structure.catalog_name = catalog_name
-        new_hc_structure.catalog_path = catalog_base_dir
-        new_hc_structure.catalog_base_dir = file_io.get_file_pointer_from_path(catalog_base_dir)
-        new_hc_structure.on_disk = True
-        # Update information on the catalog info object
-        new_hc_structure.catalog_info = dataclasses.replace(
-            new_hc_structure.catalog_info, catalog_name=catalog_name, **kwargs
-        )
-        return new_hc_structure
-
     def to_hipscat(
         self,
         base_catalog_path: str,
@@ -329,65 +258,6 @@ class Catalog(Dataset):
         Args:
             base_catalog_path (str): Location where catalog is saved to
             catalog_name (str): The name of the catalog to be saved
-            storage_options (dict): dictionary that contains abstract filesystem credentials
+            storage_options (dict): Dictionary that contains abstract filesystem credentials
         """
-        os.makedirs(base_catalog_path)
-
-        base_catalog_dir_fp = hc.io.get_file_pointer_from_path(base_catalog_path)
-
-        # Write partition parquet files
-        pixel_to_partition_size_map = self.write_partitions(base_catalog_dir_fp, storage_options)
-
-        # Write parquet metadata
-        hc.io.write_parquet_metadata(base_catalog_path, storage_options)
-
-        # Write partition info
-        partition_info = self._get_partition_info_dict(pixel_to_partition_size_map)
-        hc.io.write_partition_info(base_catalog_dir_fp, partition_info, storage_options)
-
-        # Generate most recent HiPSCat catalog information
-        new_hc_structure = self.create_modified_hc_structure(
-            base_catalog_path,
-            catalog_name if catalog_name else self.hc_structure.catalog_name,
-            total_rows=sum(pi[0] for pi in partition_info.values()),
-        )
-
-        # Write the catalog info
-        hc.io.write_catalog_info(
-            catalog_base_dir=base_catalog_path,
-            dataset_info=new_hc_structure.catalog_info,
-            storage_options=storage_options,
-        )
-
-        # Write provenance info
-        hc.io.write_metadata.write_provenance_info(
-            catalog_base_dir=base_catalog_dir_fp,
-            dataset_info=new_hc_structure.catalog_info,
-            tool_args=self._get_provenance_info(new_hc_structure),
-            storage_options=storage_options,
-        )
-
-    @staticmethod
-    def _get_provenance_info(catalog_structure) -> dict:
-        """Fill all known information in a dictionary for provenance tracking.
-
-        Returns:
-            dictionary with all argument_name -> argument_value as key -> value pairs.
-        """
-        catalog_info = catalog_structure.catalog_info
-        args = {
-            "catalog_name": catalog_structure.catalog_name,
-            "output_path": catalog_structure.catalog_path,
-            "output_catalog_name": catalog_structure.catalog_name,
-            "catalog_path": catalog_structure.catalog_path,
-            "epoch": catalog_info.epoch,
-            "catalog_type": catalog_info.catalog_type,
-            "ra_column": catalog_info.ra_column,
-            "dec_column": catalog_info.dec_column,
-        }
-        provenance_info = {
-            "tool_name": "lsdb",
-            "version": version("lsdb"),
-            "runtime_args": args,
-        }
-        return provenance_info
+        io.to_hipscat(self, base_catalog_path, catalog_name, storage_options)
