+++ src/lsdb/io/__init__.py
@@ -1,0 +1,1 @@
+from .to_hipscat import to_hipscat
